collaboration: draw both directions if one is above the top 50 threshold
In order to make the overview collaboration chart comprehensible we
visualize the top 50 connections only. A connection has two directions,
though. If the connection in one direction passes the threshold then
the one in the other direction might not. As a result only one direction
would be visualized. Visualizing the additional direction would not
introduce any additional clutter as we draw the connection already.

Fix that by visualizing both directions if at least one direction is
above the top 50 threshold.