Improved user interaction with chart tooltips.
The time series chart tooltips were a bit clunky. The area in which the
cursor had to be to invoke the tooltips was very small, and the tooltips
showed only one data entry at a time, making comparisons hard.

This commit changes that by showing tooltips already when the cursor
hovers a date (hitting the exact data points is not necessary anymore),
with all the data points available for that specific x position.