collaboration chart: draw choosen org first, make first org "blue"
If the collaboration chart of a single organization is visualized, then
ensure that the choosen org is always drawn first. This way the choosen
org is always drawn with the same color.

In addition, rotate the color wheel to start with the color "blue"
because I think it looks good.