Hyphenate within info boxes only
Previously, hyphenation was enabled for all HTML elements in the
dashboard. However, this is not very helpful with organization names, as
it’s important to know whether the organization name contains an extra
hyphen or not.

For this reason, this commit enables hyphenation for info boxes only,
for which hyphenation was originally intended, as the text columns
within info boxes are rather small.