Add anchors to items in first table columns
This adds anchors to items in the first columns of tables. This is
useful, for instance, to send links directly referencing a row in the
tables (where the first column usually some sort of index).

To disambiguate the anchor IDs in the case that there are multiple
tables on the same page, anchors are automatically prefixed with the
table’s own ID.