make "excludedEntities" a list and add "github-enterprise" as default
"github-enterprise" is an entry in the organizations table but no real
organization in GitHub Enterprise. Refactor the "excludedEntities"
mechanism to accept a list of entities instead of a single string and
exclude "github-enterprise" in the example config.

Please note that this approach also excludes repositories with the name
"github-enterprise". That shouldn't be a big problem in the real world,
though.