Fix Git download script (#46)
Let the Git download script check not only for log entries in the
previous Git audit log but also the current one. Because of the unwanted
log rotation change introduced in GitHub Enterprise 2.11, we can’t
assume that the log files searched for with .1* contain yesterday’s
information (it might still reside in the current, not-yet-rotated log).

Also remove the known issue concerning the Git download report on
GitHub Enterprise 2.11 in the hope that this bug is now fixed.