git-download: fix audit log grepping workaround for GHE 2.11+
In 50743c8 we implemented a workaround for the GitHub audit log rotation
bug. The solution was to find the log lines from yesterday in the files
that contain all logs from the last week. To achieve this we grep for
yesterday's date.

The day component of the date was padded with zeros (e.g. "Dec 04").
However, in the GitHub audit log the date is padded with spaces (e.g.
"Dec  4"). Consequently, the grep did not match for single digit days.

Fix that by padding the day with spaces.