Switch updater to systemd system service (#72)
This makes the systemd system service a user service. There are two
reasons that motivated this change. First, user services require an
active session of the specified user. This has the drawback that the
updater would not be triggered after reboots if no SSH connection to the
GitHub Enterprise appliance was made. Second, user services cannot
depend on system services. Because of that, the updater service could
not reliably wait for the rest of GitHub Enterprise to start.