Fix dh_make call for Debian packaging (#73)
dh_make recently changed from Python 2 to Python 3. For this reason, the
call to dh_make with Python 2 won’t work on systems with new dh_make
versions. Instead of locking the Python version to Python 2, call Python
without the fixed version number. This should be safe, because we can
expect Linux distributions to ship a version of dh_make that works with
the system’s default Python version.