Refactor DOM creation
This refactoring simplifies the HTML necessary to build charts by moving
the creation of required wrappers to the JavaScript side.

This approach has multiple benefits. First, the HTML part is more
lightweight. Hence, future additions to the charts (such as action
bars) will require much less changes to the HTML. Second, we have more
control over the layout on the JavaScript side without having to touch
the HTML part. Third, information concerning the chart content is
grouped and separated from the layout part more nicely in this way.
Finally, with this preparation, we could render placeholders more nicely
in the future, as they have a distinct style now.

The existing charts are adjusted to these changes and produce pixel-wise
identical visualizations. Additionally, this unifies the way how charts,
tables, and collaboration graphs are rendered.