Add action bar with link to raw data to all charts
This adds an action bar to every chart with a button to download the raw
data files in TSV format if existent. The action bar is designed to hold
more actions in the future, most notably, a switch for the level of
detail in the charts.

Along with the action bar, this commit introduces CSS for buttons.

Note that the action bar is to be considered as a “version 1” design to
bring the functionality first, and the layout will probably change in a
future release.