Make the “raw data” button optional
This introduces a new configuration option “showRawDataLink,” through
which the “raw data” button can be explicitly hidden.

This is used for hiding the “raw data” download button in the
collaboration chart, where the raw data isn’t useful to view or edit in
a spreadsheet application.

Additionally, the documentation is extended to cover the new option.