Use GitHub build hash as parameter on assets
By using the GitHub Pages build hash as a “version” parameter on the
asset URLs, we can force browser cache flushes of our JavaScript and CSS
resources upon every commit.

While it wouldn’t be strictly necessary to change the version tag of a
file on every commit (but only when it changes), this avoids having to
manually change the version parameter after each change.