Support multiple views per chart
This commit implements the new option “views.” With this array, multiple
views can be specified, which each have a unique configuration. For
instance, one view could show the daily data of the last two months and
another one weekly aggregated data for the last two years.

The implementation ensures that former configurations that don’t make
use of views still work as before.