Add view switcher to action bar
With this change, all history charts that provide multiple views in their
configurations automatically get a view switcher in the corresponding
action bar next to the chart.

This places the view switcher on the left side of the action bar, while
the data download button is moved to the right side.