PR usage: align query for active repos
Align query for active repos and add a comment to keep them aligned.
In the future we might want to refactor the "active repo" code into a
reusable function.

Although the updated query excludes entities, I don't expect the data
to change in any significant way as there is usually no activity on
the excluded entities anyways.