Unconditionally show action bar
With the view switcher added, the action bar doesn’t just consist of a
button to download the raw data now. For this reason, the action bar
shouldn’t be hidden just because the raw data button is hidden.

While this will show the action bar even if it is empty, we currently
don’t have a use case where both the view switcher and the download
button are hidden at the same time.