Make chord charts multiview-capable
This extends the chord chart with support for multiple views. This will
be used to visualize collaboration across multiple time ranges. Users
can switch between these with the view switchers that were previously
implemented for history charts.

For this to work, the data file may contain multiple matrices in the
same file (one after the other). If the configuration contains multiple
entries in the “views” array, these will be mapped to the respective
data matrices.

With this modification, the chord chart automatically generates a view
switcher and an entity (for instance, organization) selector. The
visualization is constantly kept in sync with these inputs.