Fix error with empty chart views (#134)
When aggregating over only few data points, the weekly-aggregated views
might end up empty. This led to an error, because the chart generation
relied on the data not to be empty.

This adds a corresponding check, which prevents the error and shows an
empty chart as expected.

Instead of showing users empty charts, views with no data should be
hidden from the users. This will be addressed separately.