Rewrite headers on every update (#138)
With this commit, every single update will cause the data file headers
to be rewritten to their defaults.

In this way, we can easily change column labels in the data files in the
future without having to worry about empty charts due to the inconsistent
naming scheme (which causes the dashboard not to match the data series
with the columns appropriately).

I don’t consider this a change that could accidentally break data by,
for instance, switching the meaning of columns in the future. The reason
is that we have the data schema mechanism in place to deal with cases
where we would like to make deeper changes to the data format that
would require a data migration.