Fix tokenless authentication chart
The format of the gitauth.log file changed from GitHub Enterprise 2.11
to 2.12. This makes it necessary to adjust the regular expression
responsible for fetching user names and repositories.

Unfortunately, it seems that we can no longer rely on the order of the
fields in gitauth.log. For this reason, multiple grep iterations are
performed and the regular expression is changed to use look-aheads to
allow for arbitrary field orders.

For backward compatibility with GitHub Enterprise 2.11 and earlier, a
switch is implemented that executes the old query if necessary.