Improve tooltips when aggregating data
This implements multiple improvements to the tooltips that are shown
when hovering data points in views displaying aggregated data.

First, the tooltips now show the date range of the period of time over
which the data has been aggregated (and not just the first date of the
range). An exception to this are the “first” and “last” aggregation
methods (which only pick the first or last value from the available time
range), where the respective date is shown instead.

For reasons of clarity, the date format was changed to show abbreviated
month names instead of full ones.

Second, the applied aggregation method is shown in parentheses next to
the date range. This makes it easier to understand how to interpret the
data. Again, this excludes the “first” and “last” aggregation methods,
where the date itself already provides all the information needed.

Implementation-wise, new auxiliary columns were introduced. These are
distinguished from regular columns with a leading underscore character
and subsequently ignored from the chart rendering. This mechanism is
useful for annotating data, currently with information for showing the
tooltips.