Add detailed Git versions report
This adds an extended edition of the Git versions report. In addition to
tracking the used Git versions of the last day, this tracks all used Git
versions every day, making an upgrade trend chart possible.

Note that this only includes the report (which gathers the data) but not
the chart. Also, for compatibility, the former Git versions chart is
retained for now. The migration will be done in a future pull request.