added meta.tsv, updated README.md, fixed calendar import
added an example meta.tsv file that can be used for the hubble-data
repository. Update the README.md to tell people that a meta.tsv file
has to be present inside hubble-data. Added an Exception and print
output in case there is no meta.tsv file. Fixed missing inmport statement
in ReportPRUsage.py for callendar