Support empty data repositories
By accident, newly created data repositories didn’t work with the
updater. The reason is that the updater checked the data schema version
but failed, because no meta.tsv file (which contains the data schema
version) was present.

This adds a check that, if the meta.tsv file is not yet present, simply
succeeds with the update, through which the data repository is populated
with an up-to-date schema anyway.