Show median values in aggregated charts by default
This switches all charts with the “first” aggregation strategy to
“median” aggregation.

Previously, chart views with aggregated data just showed the first value
of each time period. The median value might be a better choice, because
it is more representative of the data than just picking the first value
available. The arithmetic mean was also considered as a choice, but is
expected not to even out sporadic effects such as public holidays as
nicely as the median function.