Decompose charts.js
The JavaScript part of the dashboard is contained in a single charts.js
file that has grown over time and that became quite hard to read,
maintain, or extend.

This refactoring uses the “include_relative” Jekyll directive, which can
safely be used on GitHub Pages, because it is part of the “gh_pages”
gem. This directive allows assets (not just pages) to be composed of
partial source files.

This commit splits charts.js into special-purpose partial JavaScript
files containing shared utilities, style-related definitions, as well as
the charts themselves.

Unfortunately, this approach breaks our testing setup, the reason being
that the testing framework looks at the unprocessed source files, and
the “include_relative” directives aren’t valid JavaScript code. To work
around this issue, this commit first invokes Jekyll to generate the
final charts.js from the partials before running the tests.