Add JSON file with information about Git releases
The “git-versions.json” file contains information about the current Git
releases. This includes all released stable versions of Git along with
their release dates as well as a list of all public vulnerabilities to
Git and which versions they affect.

The intent of this file is to be able to discern which Git versions are
not recommended for productive use or affected by a known vulnerability.

This file is up-to-date as of 2018-10-08, including the very recently
disclosed CVE-2018-17456 vulnerability.

Vulnerabilities in versions released before 2017 were omitted for
simplicity, because all Git versions released before 2017 are affected
by at least one vulnerability.