Implement recommended Git versions chart
This adds a chart that shows how many users are using Git clients that
are recommended, outdated, and vulnerable as a stacked area chart.

The chart obtains its data from the previously introduced Git release
database and uses an area chart for visualizing the results. The actual
recommended Git versions chart is built on top of the area chart and
performs only the operations necessary to prepare the data accordingly.

The unit tests cover the correct computation of how many versions were
recommended, outdated, and vulnerable at a given date based on recorded
usage data.