Adjust colors in recommended Git versions chart
This adds an option to specify custom colors for datasets in the area
chart. This is done through a mapping from dataset IDs to color
specifiers chosen from a built-in list of color names.

With this option, the recommended Git versions chart is colorized such
that vulnerable versions are rendered red, outdated ones yellow, and
recommended ones green.