Load online Git release database if accessible
With this patch, the Git release database is preferentially retrieved
online from the official Hubble demo page [1] if it is accessible.

This has the advantage that users will see more recent Git version
information than is locally available on their local installation
(which dates back to the respective release date).

The idea is to regularly update the Git release database in the public
repository to make recent information about vulnerabilities immediately
visible to Hubble users.

In the case that the remote copy of the Git release database is not
accessible (for instance, because of corporate network restrictions),
the local copy will be used as a fallback.

[1] https://autodesk.github.io/hubble/assets/js/git-versions.json