read Git client versions from audit log
Previously we read the versions from the haproxy log and used the
originating IP address to distinguish machines/users. In a load balancer
setup this IP address can be the same for all requests and consequently
the number of users per version are not counted correctly.

Fix this by using the audit log to count the Git versions. Also count
the versions by user ID and not by IP address.

see https://github.com/Autodesk/hubble/issues/184