Update Git version database
This brings the Git version database up-to-date with the Git 2.26.1
release.

Starting with this commit, the publication date of Git releases does not
correspond with the Git release tag time stamps anymore. This proved to
be problematic because the Git maintainers sometimes decide to hold back
releases with security fixes for up to multiple weeks, even though the
release tags have already been made.

Instead, publication time stamps are now manually taken from the
announcements on the Git mailing list.