Fix date in unit test
The Git versions database was recently updated to reflect the actual
publication date of Git releases (which may or or may not coincide with
the tag time stamp of the respective release). This updates the
timestamp in the unit tests for the Git versions database accordingly,
as it would incorrectly fail otherwise.