Fix license identifier in Debian packaging
This fixes an issue with Debian packaging, where the license identifier
“mit” isn’t accepted anymore. This replaces it with “expat,” which
refers to the exact same license, the different name only considered
necessary to clearly identify this license.

Some argue that the term “MIT License” is ambiguous, as it may refer to
two different versions of the license text. (Both of them only have a
very subtle difference though, where X Consortium trademarks may be
mentioned or not.) However, the Debian folks apparently now dropped the
“mit” identifier entirely, forcing users to choose the more specific
identifier “expat” instead or “custom” if it doesn’t fit.