Refactor functions for server calls
The function executeScript used to have a double role, as it was in
charge of executing both Unix commands on the server (locally or by
using a local connection) but also Bash scripts.

This was confusing per se, but had a few side effects that might have
been unexpected to developers looking at the code.

First of all, the optional stdin argument was completely ignored when
executing Bash scripts on the server. The reason for that is that stdin
was used to pass the script’s content to the server (using the -s option
provided by Bash). In a sense, the interface of this function was
misleading as a result of this.

Also, when using executeScript to run regular Unix commands, unnecessary
file look-ups were made.

To address these issues, this replaces the executeScript method with two
new ones: executeCommandOnServer (which executes regular Unix commands)
and executeBashScriptOnServer (which executes Bash scripts and currently
does not accept stdin because of the limitation mentioned above).

Co-authored-by: Marius Knaust <marius.knaust@gmail.com>