Separate command arguments from name
In order to execute commands, subprocess.Popen expects a list of
strings, where the first one represents the name of the command and the
remaining strings are the command’s arguments. However, the -s option
was included in the name of the binary, which isn’t supported in newer
versions of	Python anymore.

Co-authored-by: Milos Svasek <milos@svasek.net>