Fix Ruby script execution
Ruby scripts were executed on the server by passing them to the GitHub
environment runner as a command-line argument. This caused trouble with
newlines, which were escaped with additional backslashes. However, it
appears that this escaping may not be correct, depending on whether the
command is run over SSH or not.

With this change, the Ruby script is not provided as a command-line
argument anymore. Instead, the GitHub environment runner is instructed
to read the Ruby script from stdin. This solution has the benefit of
being more robust concerning newlines, as those don’t need to be escaped
anymore. Also, this solution works regardless of other potentially
confounding factors such as running the updater in a systemd environment
vs. a remote shell or subprocess implementation differences between
Python versions.